[py3] Re-re-do of the charDataElement in bitwise
So, I've been a little bit on unicode autopilot, just trying to get the
data types to match up. However, once getting into the work of making
the drivers work, we have a ton that do things like this:

  memobj.name = 'foobar\xFF'

because the radio needs the \xFF terminating byte to know where the end
of the string is. We need to not encode/decode that even on py3 because
we won't end up with the right value in memory if we do (and we'll fail
to encode/decode some values).

So, this cheats our unicodeness to end up with single-byte characters and
the right things get into/out of memory for now. Ideally all drivers would
be converted to bytestrings for things like that, but, babysteps.

Yes, I know how unicode works in Python, and yes I know this is not how
it's supposed to work.

#495

--HG--
branch : py3