[py3] A workaround for match_model legacy, and test harness updates
This is enough to run the driver tests in py3 to actually start hitting issues.
The filestring hack in directory just lets us get to some of them without
having to convert match_model on all of them. Same deal with skipping drivers
that don't load if CHIRP_DEBUG is set.

#495

--HG--
branch : py3